+++ src/test/java/lt/management/oms/OrderManagementSystemApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class OrderManagementSystenApplicationTests {
+class OrderManagementSystemApplicationTests {
 
 	@Test
 	void contextLoads() {
